Fix spaCy install with GPU support
Need to install both spaCy and spacy-transformers with cuda support.
Don't lock to cuda 10.0, since the Docker image has 10.1.  To avoid forgetting
to update this and breaking it in the future, we'll accept a longer
build time and install the generic cuda version that requires compilation
instead of a hardcoded version with a wheel.